+++ mhs/outbound/Pipfile
@@ -11,7 +11,7 @@ unittest-xml-reporting = "~=2.5"
 requests = "*"
 tornado = "*"
 defusedxml = "~=0.6"
-mhs-common = {editable = true,path = "./../mhs_common"}
+mhs-common = {editable = true,path = "./../common"}
 integration-adaptors-common = {editable = true,path = "./../../common"}
 
 [requires]
