+++ mhs/outbound/outbound/transmission/tests/test_outbound_transmission.py
@@ -4,7 +4,7 @@ from unittest.mock import patch, sentinel, Mock
 
 import requests
 
-from outbound import transmission as outbound_transmission
+from outbound.transmission import outbound_transmission
 
 URL_NAME = "url"
 URL_VALUE = "URL"
